// src/proto/tyto/runtime/v1/guest.ts
export enum FileKind {
  FILE_KIND_UNSPECIFIED = 0,
  FILE_KIND_FILE = 1,
  FILE_KIND_DIRECTORY = 2,
  FILE_KIND_SYMLINK = 3,
  FILE_KIND_OTHER = 4,
}
// ...
export interface FileInfo {
  path: string;
  name: string;
  kind: FileKind;
  size: number;
  mode: number;
  modifiedAtUnixNanos: number;
}
// ...
export const FileInfo: MessageFns<FileInfo> = {
  encode(message: FileInfo, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.path !== "") {
      writer.uint32(10).string(message.path);
    }
    if (message.name !== "") {
      writer.uint32(18).string(message.name);
    }
    if (message.kind !== 0) {
      writer.uint32(24).int32(message.kind);
    }
    if (message.size !== 0) {
      writer.uint32(32).uint64(message.size);
    }
    if (message.mode !== 0) {
      writer.uint32(40).uint32(message.mode);
    }
    if (message.modifiedAtUnixNanos !== 0) {
      writer.uint32(48).int64(message.modifiedAtUnixNanos);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FileInfo {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFileInfo();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.path = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.kind = reader.int32() as any;
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.size = longToNumber(reader.uint64());
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.mode = reader.uint32();
          continue;
        }
        case 6: {
          if (tag !== 48) {
            break;
          }

          message.modifiedAtUnixNanos = longToNumber(reader.int64());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FileInfo {
    return {
      path: isSet(object.path) ? globalThis.String(object.path) : "",
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      kind: isSet(object.kind) ? fileKindFromJSON(object.kind) : 0,
      size: isSet(object.size) ? globalThis.Number(object.size) : 0,
      mode: isSet(object.mode) ? globalThis.Number(object.mode) : 0,
      modifiedAtUnixNanos: isSet(object.modifiedAtUnixNanos)
        ? globalThis.Number(object.modifiedAtUnixNanos)
        : isSet(object.modified_at_unix_nanos)
        ? globalThis.Number(object.modified_at_unix_nanos)
        : 0,
    };
  },

  toJSON(message: FileInfo): unknown {
    const obj: any = {};
    if (message.path !== "") {
      obj.path = message.path;
    }
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.kind !== 0) {
      obj.kind = fileKindToJSON(message.kind);
    }
    if (message.size !== 0) {
      obj.size = Math.round(message.size);
    }
    if (message.mode !== 0) {
      obj.mode = Math.round(message.mode);
    }
    if (message.modifiedAtUnixNanos !== 0) {
      obj.modifiedAtUnixNanos = Math.round(message.modifiedAtUnixNanos);
    }
    return obj;
  },

  create(base?: DeepPartial<FileInfo>): FileInfo {
    return FileInfo.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<FileInfo>): FileInfo {
    const message = createBaseFileInfo();
    message.path = object.path ?? "";
    message.name = object.name ?? "";
    message.kind = object.kind ?? 0;
    message.size = object.size ?? 0;
    message.mode = object.mode ?? 0;
    message.modifiedAtUnixNanos = object.modifiedAtUnixNanos ?? 0;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}